requirements.txt
aiosqlite==0.17.0
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
atlastk==0.13.3
certifi==2023.7.22
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.103.1
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
iso8601==1.1.0
pydantic==2.3.0
pydantic_core==2.6.3
pypika-tortoise==0.1.6
python-decouple==3.8
pytz==2023.3.post1
sniffio==1.3.0
starlette==0.27.0
tortoise==0.1.1
tortoise-orm==0.20.0
typing_extensions==4.7.1
uvicorn==0.23.2
